requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
plotly>=5.19.0
scikit-learn>=1.3.0
ta>=0.10.0
python-jose>=3.3.0
bcrypt>=4.1.0
TA-Lib>=0.4.24
PyJWT>=2.8.0
aiohttp>=3.9.0
